Separate into library and binary
