+++ bin/boxes/errorbox.rs
