+++ bin/boxes/linkbox.rs
