+++ bin/pages/file_explorer.rs
@@ -13,7 +13,7 @@ use ratatui::{
 };
 
 use crate::search::find_files;
-use crate::util::colors::COLOR_CONFIG;
+use md_tui::config::colors::COLOR_CONFIG;
 
 #[derive(Debug, Clone)]
 enum MdFileComponent {
@@ -100,6 +100,7 @@ impl FileTree {
         this
     }
 
+    #[allow(dead_code)]
     pub fn sort(&mut self) {
         let filtered: Vec<&MdFile> = self
             .files
@@ -287,6 +288,7 @@ impl FileTree {
         self.files.push(MdFileComponent::Spacer);
     }
 
+    #[allow(dead_code)]
     pub fn files(&self) -> Vec<&MdFile> {
         self.files
             .iter()
@@ -297,6 +299,7 @@ impl FileTree {
             .collect::<Vec<&MdFile>>()
     }
 
+    #[allow(dead_code)]
     pub fn all_files(&self) -> &Vec<MdFile> {
         &self.all_files
     }
