+++ bin/util/keys.rs
