+++ src/config.rs
@@ -1,0 +1,2 @@
+pub mod colors;
+pub mod general;
