--- src/pages/mod.rs
@@ -1,2 +1,0 @@
-pub mod file_explorer;
-pub mod markdown_renderer;
